feat: Add benchmarks, unit tests for audio analyser and refactor of gui

+++ src/core/AudioAnalyser/AudioAnalyser.hpp
@@ -22,15 +22,15 @@ public:
      */
     struct FrequencyData
     {
-        float amplitude;
-        float phase;
-        float frequency;
+        double amplitude;
+        double phase;
+        double frequency;
     };
 
     AudioAnalyser();
 
     /**
-     * @brief Decomposes a time-domain audio signal into its frequency components using the Discrete Fourier Transform (DFT).
+     * @brief Computes the frequency spectrum of a signal using the Fast Fourier Transform (DFT).
      *
      * @details
      * The Discrete Fourier Transform converts a sequence of N time-domain samples into N frequency-domain components.
@@ -48,13 +48,13 @@ public:
      *
      * The computational complexity of this naive DFT implementation is O(N^2), where N is the number of samples.
      *
-     * @param samples A vector of float values representing the audio samples in the time domain.
+     * @param samples A vector of double values representing the audio samples in the time domain.
      * @param sampleRate The sampling rate of the audio signal in Hz.
      *
      * @return std::vector<FrequencyData> A vector containing the frequency components, each with amplitude, phase, and frequency information.
      */
     static std::vector<FrequencyData>
-    discreteFourierTransform(const std::vector<float> &samples, int sampleRate);
+    discreteFourierTransform(const std::vector<double> &samples, int sampleRate);
 
     /**
      * @brief Reconstructs a time-domain audio signal from its frequency components.
@@ -75,9 +75,9 @@ public:
      * @param numSamples The number of samples to generate in the reconstructed signal.
      * @param sampleRate The sampling rate of the audio signal in Hz.
      *
-     * @return std::vector<float> A vector of float values representing the reconstructed audio samples in the time domain.
+     * @return std::vector<double> A vector of double values representing the reconstructed audio samples in the time domain.
      */
-    static std::vector<float>
+    static std::vector<double>
     reconstructDFT(const std::vector<FrequencyData> &spectrum, int numSamples, int sampleRate);
 
     /**
@@ -105,13 +105,64 @@ public:
      *
      * The results of recursive calls are combined using the roots of unity to produce the final transformed vector.
      *
+     * @par Polynomial decomposition (Cooley–Tukey FFT)
+     *
+     * The polynomial in coefficient form:
+     * \f[
+     * P(x) = c_0 + c_1 x + c_2 x^2 + \dots + c_{n-1} x^{n-1}
+     * \f]
+     *
+     * can be split into even and odd coefficients:
+     * \f[
+     * P(x) = \underbrace{(c_0 + c_2 x^2 + c_4 x^4 + \dots)}_{P_e(x^2)}
+     * + x \cdot \underbrace{(c_1 + c_3 x^2 + c_5 x^4 + \dots)}_{P_o(x^2)}
+     * \f]
+     *
+     * which gives:
+     * \f[
+     * P(x) = P_e(x^2) + x \cdot P_o(x^2)
+     * \f]
+     *
+     * Evaluating at roots of unity \f$x_j\f$ and \f$-x_j\f$:
+     * \f[
+     * P(x_j) = P_e(x_j^2) + x_j \cdot P_o(x_j^2)
+     * \f]
+     * \f[
+     * P(-x_j) = P_e(x_j^2) - x_j \cdot P_o(x_j^2)
+     * \f]
+     *
+     * Substituting \f$x_j = \omega^j\f$ (where \f$\omega\f$ is a primitive root of unity):
+     * \f[
+     * P(\omega^j) = P_e(\omega^{2j}) + \omega^j \cdot P_o(\omega^{2j})
+     * \f]
+     * \f[
+     * P(\omega^{j + n/2}) = P_e(\omega^{2j}) - \omega^j \cdot P_o(\omega^{2j})
+     * \f]
+     *
+     * Defining:
+     * \f[
+     * y_{\text{even}}[j] = P_e(\omega^{2j}), \quad
+     * y_{\text{odd}}[j] = P_o(\omega^{2j})
+     * \f]
+     *
+     * we obtain the final FFT combination step:
+     * \f[
+     * y[j] = y_{\text{even}}[j] + \omega^j \cdot y_{\text{odd}}[j]
+     * \f]
+     * \f[
+     * y[j + n/2] = y_{\text{even}}[j] - \omega^j \cdot y_{\text{odd}}[j]
+     * \f]
+     *
+     * These equations are directly used in the FFT implementation.
+     *
+     *
      * @param polynomial Vector of complex coefficients representing the polynomial.
      * @param invert Boolean flag indicating whether to perform inverse transform (true) or forward transform (false).
      *
-     * @return std::vector<std::complex<float>> The transformed polynomial in value or coefficient form depending on invert flag.
+     * @return std::vector<std::complex<double>> The transformed polynomial in value or coefficient form depending on invert flag.
      */
-    static std::vector<std::complex<float>>
-    fftRecursive(std::vector<std::complex<float>> &polynomial, bool invert);
+    static std::vector<std::complex<double>>
+    fftRecursive(std::vector<std::complex<double>> &polynomial, bool invert);
 
     /**
      * @brief Computes the frequency spectrum of a signal using the Fast Fourier Transform (FFT).
@@ -125,13 +176,13 @@ public:
      *
      * This method is preferred over the naive DFT for large input sizes due to its significant performance improvement.
      *
-     * @param samples Vector of float audio samples in the time domain, size must be a power of two.
+     * @param samples Vector of double audio samples in the time domain, size must be a power of two.
      * @param sampleRate Sampling rate of the audio signal in Hz.
      *
      * @return std::vector<FrequencyData> Vector of frequency components with amplitude, phase, and frequency.
      */
     static std::vector<FrequencyData>
-    fastFourierTransform(const std::vector<float> &samples, int sampleRate);
+    fastFourierTransform(const std::vector<double> &samples, int sampleRate);
 
     /**
      * @brief Reconstructs a time-domain signal from its FFT-derived frequency spectrum.
@@ -153,9 +204,9 @@ public:
      * @param numSamples Number of samples to generate in the reconstructed time-domain signal.
      * @param sampleRate Sampling rate of the audio signal in Hz.
      *
-     * @return std::vector<float> Vector of reconstructed audio samples in the time domain.
+     * @return std::vector<double> Vector of reconstructed audio samples in the time domain.
      */
-    static std::vector<float>
+    static std::vector<double>
     reconstructFFT(const std::vector<FrequencyData> &spectrum, int numSamples, int sampleRate);
 
 private:
