+++ src/gui/widgets/OverviewWidget/OverviewWidget.cpp
@@ -16,7 +16,7 @@ OverviewWidget::OverviewWidget(QWidget *parent) : QWidget(parent)
 
 void OverviewWidget::displayData(const AudioUtils::AudioData &audioData,
                                  const std::vector<AudioAnalyser::FrequencyData> &spectrum,
-                                 const std::vector<float> &reconstructedSamples)
+                                 const std::vector<double> &reconstructedSamples)
 {
     QLayoutItem *item;
     while ((item = layout->takeAt(0)) != nullptr)
